Fixing index and linter
